refactor first draft
