+++ src/go.mod
@@ -1,0 +1,14 @@
+module github.com/d3tourrr/NomiKin-Discord
+
+go 1.18
+
+require (
+	github.com/bwmarrin/discordgo v0.28.1
+	github.com/d3tourrr/NomiKinGo v0.2.41
+)
+
+require (
+	github.com/gorilla/websocket v1.5.3 // indirect
+	golang.org/x/crypto v0.29.0 // indirect
+	golang.org/x/sys v0.28.0 // indirect
+)
