+++ src/messagequeue.go
@@ -1,0 +1,56 @@
+package main
+
+import (
+    "log"
+    "sync"
+    "time"
+
+    "github.com/bwmarrin/discordgo"
+)
+
+type QueuedMessage struct {
+    Message   *discordgo.MessageCreate
+    Companion *Companion
+}
+
+type MessageQueue struct {
+    messages []QueuedMessage
+    mu       sync.Mutex
+}
+
+func (q *MessageQueue) Enqueue(message QueuedMessage) {
+    q.mu.Lock()
+    defer q.mu.Unlock()
+    q.messages = append(q.messages, message)
+}
+
+func (q *MessageQueue) Dequeue() (QueuedMessage, bool) {
+    q.mu.Lock()
+    defer q.mu.Unlock()
+
+    if len(q.messages) == 0 {
+        return QueuedMessage{}, false
+    }
+
+    message := q.messages[0]
+    q.messages = q.messages[1:]
+    return message, true
+}
+
+func (q *MessageQueue) ProcessMessages() {
+    for {
+        queuedMessage, ok := q.Dequeue()
+        if !ok {
+            time.Sleep(1 * time.Second) // No messages in queue, sleep for a while
+            continue
+        }
+
+        err := SendMessageToCompanion(queuedMessage.Message, queuedMessage.Companion)
+        if err != nil {
+            log.Printf("Failed to send message to Companion API: %v", err)
+            q.Enqueue(queuedMessage) // Requeue the message if failed
+        }
+
+        time.Sleep(5 * time.Second) // Try to keep from sending messages toooo quickly
+    }
+}
